requirements/production.txt
Django
psycopg2
prettyconf
dj-database-url
djangorestframework
coreapi
markdown
pygments
